Fix lint and type errors
