+++ lighthouse-plugin-ad-speed-insights/gatherers/async-call-stacks.js
@@ -1,10 +1,18 @@
-/**
- * @license Copyright 2018 Google Inc. All Rights Reserved.
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License. You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
- * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
- */
-'use strict';
+// Copyright 2019 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
 
+// @ts-ignore
 const Gatherer = require('lighthouse/lighthouse-core/gather/gatherers/gatherer.js');
 
 /**
@@ -15,14 +23,12 @@ class AsyncCallStacks extends Gatherer {
    * @param {LH.Gatherer.PassContext} passContext
    */
   async beforePass(passContext) {
-    for (const _ of [0, 0]) {
-      await passContext.driver.sendCommand(
-        'Runtime.setAsyncCallStackDepth', {maxDepth: 1000});
-      await new Promise((resolve) => setTimeout(resolve, 200));
-      await passContext.driver.sendCommand(
-        'Debugger.setAsyncCallStackDepth', {maxDepth: 2000});
-      await new Promise((resolve) => setTimeout(resolve, 200));
-    }
+    await passContext.driver.sendCommand(
+      'Runtime.setAsyncCallStackDepth', {maxDepth: 1000});
+    await new Promise((resolve) => setTimeout(resolve, 100));
+    await passContext.driver.sendCommand(
+      'Debugger.setAsyncCallStackDepth', {maxDepth: 1000});
+    await new Promise((resolve) => setTimeout(resolve, 100));
   }
 
   /**
